update window enumerator

+++ AutoWindowCapture/AutoWindowCapture/Services/RecordingWindowLauncher.cs
@@ -1,13 +1,13 @@
 using System;
 using System.Collections.Generic;
-using WindowEnumerator;
+using WindowEnumeratorNative;
 using WindowListPage.Services;
 
 namespace AutoWindowCapture.Services;
 
 public sealed class RecordingWindowLauncher : IRecordingWindowLauncher
 {
-    private readonly Dictionary<IntPtr, RecordingWindow.Views.RecordingWindow> _windowsByHandle = [];
+    private readonly Dictionary<long, RecordingWindow.Views.RecordingWindow> _windowsByHandle = [];
 
     public void OpenOrActivate(WindowInfo windowInfo, bool recOnOpen)
     {
